chore(subgraph): pull trait impl into seperate file for clarity
