+++ crates/subgraph/src/lib.rs
@@ -1,13 +1,14 @@
-mod orderbook_client;
 mod csv;
 mod cynic_client;
+mod orderbook_client;
 mod pagination;
 pub mod types;
 pub mod utils;
+mod vault_balance_changes_query;
 
 #[cynic::schema("orderbook")]
 pub mod schema {}
 
-pub use orderbook_client::{OrderbookSubgraphClient, OrderbookSubgraphClientError};
 pub use csv::{TryIntoCsv, TryIntoCsvError};
+pub use orderbook_client::{OrderbookSubgraphClient, OrderbookSubgraphClientError};
 pub use pagination::{PageQueryClient, PaginationArgs};
