+++ crates/subgraph/src/types/mod.rs
@@ -6,6 +6,6 @@ pub mod order_take_detail;
 pub mod order_takes_list;
 pub mod orders_list;
 pub mod vault_balance_change;
-pub mod vault_detail;
 pub mod vault_balance_changes_list;
+pub mod vault_detail;
 pub mod vaults_list;
